Drop command name from args when parsing them
According to the documentation, ArgumentParser.parse_known_args() seems
to work without the command name as first argument. Since CliCommand is
parsing the parameters of a specific command, we can drop the command
itself in addition to the application name before calling
ArgumentParser.parse_known_args(); this avoids the need to drop it after
parsing.

In addition, tweak the check for missing test arguments: 'if not foo' is
also matched by an empty list, which now is a valid list of arguments to
pass for tests. Hence, change it to check for None explicitly.

Adapt the tests accordingly.

(cherry picked from commit d76511019eebf309c82ebcc78124c66d327fd3fe)