1898563: move syspurpose subcommands within the 'syspurpose' command
As an effort to consolidate the handling of all the syspurpose
attributes in a single place, make sure that all the syspurpose
attributes are handled as subcommands of the 'syspurpose' command
(leaving the old commands as deprecated).

Make AbstractSyspurposeCommand able to override the ArgumentParser
created by AbstractCLICommand (its grandparent class), so it is possible
to create a subparser instead of a standalone parser.

Then, make SyspurposeCommand create all the syspurpose subcommands, with
their parsers created as subparsers of it; properly dispatch the parsed
command line options to the right subcommand when SyspurposeCommand is
executed.

Finally, change the short description of each subcommand to say that is
deprecated when shown in the list of the subscription-command commands.

(cherry picked from commit 5406865e8b1d85716a07530274028aa4f6e418b7)