1974641: Fix tab completion with multiple optional commands
* Card ID: ENT-4013
* BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1974641

After the syspurpose was added as subcommand in 5406865 the 'prev'
variable was not always set correctly. For the first argument the
completion worked fine, but for the second (and further) the last
argument was picked up instead of the last subcommand.

This change iterates over the CLI arguments and uses the last word that
does not start with a dash, skipping all arguments and finding the last
subcommand.

(cherry picked from commit 192e98bb87034e65393cc94e2698f93ad0da45dd)