1984133: repos: respect order of --enable & --disable
With the old optparse implementation, there was a single callback to
fill 'repo_actions' with the various values of the --enable & --disable
options, in the same order they were specified on the command line.
Commit eef1bb13f8924af9539f98330465c110f2df92b5 switched to argparse,
and switched to different destination lists for --enable & --disable,
mixing them back to a single list with first all the repos to enabled
and then all the repos to disable. As result, the order of intermixed
--enable & --disable was lost.

To revert back to the wanted behaviour, implement a custom
argparse.Action for --enable & --disable that appends all the actions to
the same list in the same order they were specified. This removes the
need for post-processing later on.

Card-ID: ENT-4146
(cherry picked from commit 8ac4b0f68611a295bc140ecd39279621addefef5)